Drag-Diffusion/requirements.txt
torch>=2.0
torchvision
diffusers>=0.21
transformers>=4.35
accelerate
safetensors>=0.3.1
gradio>=4.0
Pillow
numpy
matplotlib
